+++ cmd/curator/main.go
@@ -21,7 +21,7 @@ import (
 )
 
 const (
-	version = "0.17.0"
+	version = "0.17.1"
 )
 
 func main() {